Move generateSpend() into AbstractConserveAmount
Move generateSpend() into AbstractConserveAmount and make it generic enough to use for other
similar contracts to cash, rather than just cash. Also removes duplicated logic in gathering
coins.
